C:/Program Files/Git/v1/api/receive-from-order API実装

+++ src/main/java/com/example/backend/dto/InventoryDispatchRequest.java
@@ -3,12 +3,14 @@
 package com.example.backend.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
+
 
 @Data
 public class InventoryDispatchRequest {
 
   private String itemCode; // 商品コード（必須）
-  private Integer quantity; // 出庫数（必須）
+  private BigDecimal quantity; // 出庫数（必須）
   private String operator; // 担当者（必須）
   private String remarks; // 備考（任意）
 }
