+++ src/main/java/com/example/backend/dto/InventoryReceiveFromOrderRequest.java
@@ -2,6 +2,7 @@ package com.example.backend.dto;
 
 import lombok.Data;
 import java.util.List;
+import java.math.BigDecimal;
 
 @Data
 public class InventoryReceiveFromOrderRequest {
@@ -12,7 +13,7 @@ public class InventoryReceiveFromOrderRequest {
   @Data
   public static class Item {
     private String itemCode;
-    private Integer receivedQuantity;
+    private BigDecimal receivedQuantity;
     // ↓ リクエスト側で明示的に価格を渡さず、サーバー側で purchase_order_detail から取得する方針に変更
     // private Integer purchasePrice;
     private String remarks;
