+++ src/main/java/com/example/backend/dto/InventoryReceiveRequest.java
@@ -10,7 +10,7 @@ public class InventoryReceiveRequest {
 
   private String orderNo; // 発注番号（必須）
   private String itemCode; // 商品コード（必須）
-  private Integer quantity; // 入庫数（必須）
+  private BigDecimal quantity; // 入庫数（必須）
   private String operator; // 担当者（必須）
 
   private String manufacturer; // メーカー（任意）
