+++ src/main/java/com/example/backend/dto/StockMasterRequest.java
@@ -1,6 +1,7 @@
 // 在庫マスタ登録用
 package com.example.backend.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -10,5 +11,5 @@ public class StockMasterRequest {
   private String itemName; // 名称
   private String modelNumber; // 型番（任意）
   private String category; // カテゴリ
-  private Integer currentStock; // 初期在庫（任意）
+  private BigDecimal currentStock; // 初期在庫（任意）
 }
